login fn:
- add auth context for app and small fixes
