+++ src/App.css
@@ -1,7 +1,6 @@
 #root {
   max-width: 1280px;
   margin: 0 auto;
-  padding: 2rem;
   text-align: center;
   min-height: 100vh;
   width: 100%;
