+++ src/pages/UserProfile/UserProfile.tsx
@@ -1,4 +1,4 @@
-import { useState } from 'react';
+import React, { useState } from 'react';
 import styles from './UserProfile.module.scss';
 
 export default function UserProfile() {
