+++ src/pages/auth/AuthContext/AuthContext.tsx
@@ -1,0 +1,21 @@
+import React, { createContext, Dispatch, ReactNode, SetStateAction, useState } from "react";
+
+interface AuthContextType {
+    isAuthenticated: boolean;
+    setIsAuthenticated: Dispatch<SetStateAction<boolean>>;
+    username: string | null;
+    setUsername: Dispatch<SetStateAction<string | null>>;
+}
+
+export const AuthContext = createContext<AuthContextType>({} as AuthContextType);
+
+export const AuthProvider = ({children}: { children: ReactNode }) => {
+    const [isAuthenticated, setIsAuthenticated] = useState<boolean>(false);
+    const [username, setUsername] = useState<string | null>(null);
+
+    return (
+        <AuthContext.Provider value={{isAuthenticated, setIsAuthenticated, username, setUsername}}>
+            {children}
+        </AuthContext.Provider>
+    );
+};
