+++ src/pages/auth/Login.module.scss
@@ -1,0 +1,5 @@
+.auth_form_input_container {
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+}
